refactor(teams): rename member entity to user
