+++ api/src/Teams.API/Extensions/Extensions.cs
@@ -6,8 +6,8 @@ using Teams.API.Features.Projects.CreateProject;
 using Teams.API.Features.Projects.GetProjectById;
 using Teams.API.Logging;
 using Teams.API.Validation;
-using Teams.Domain.Aggregates.MemberAggregate;
 using Teams.Domain.Aggregates.ProjectAggregate;
+using Teams.Domain.Aggregates.UserAggregate;
 using Teams.Infrastructure;
 
 namespace Teams.API.Extensions;
@@ -36,12 +36,12 @@ internal static class Extensions
             options.UseNpgsql(builder.Configuration.GetConnectionString("DefaultConnection"));
             options.UseSeeding((context, _) =>
             {
-                var member = context.Set<Member>()
+                var user = context.Set<User>()
                     .FirstOrDefault(m => m.IdentityGuid == builder.Configuration["SeedUser:IdentityGuid"]);
-                if (member == null)
+                if (user == null)
                 {
-                    member = new Member(builder.Configuration["SeedUser:IdentityGuid"]!);
-                    context.Set<Member>().Add(member);
+                    user = new User(builder.Configuration["SeedUser:IdentityGuid"]!);
+                    context.Set<User>().Add(user);
                     context.SaveChanges();
                 }
                 
@@ -52,7 +52,7 @@ internal static class Extensions
                     project = new Project(
                         "Sample Project",
                         "Sample Text.",
-                        member.Id);
+                        user.Id);
                     context.Set<Project>().Add(project);
                     context.SaveChanges();
                 }
