+++ api/src/Teams.Infrastructure/TeamDbContext.cs
@@ -1,7 +1,7 @@
 using Microsoft.EntityFrameworkCore;
-using Teams.Domain.Aggregates.MemberAggregate;
 using Teams.Domain.Aggregates.ProjectAggregate;
 using Teams.Domain.Aggregates.TeamAggregate;
+using Teams.Domain.Aggregates.UserAggregate;
 using Teams.Infrastructure.EntityConfigurations;
 
 namespace Teams.Infrastructure;
@@ -10,14 +10,14 @@ public class TeamDbContext : DbContext
 {
     public TeamDbContext(DbContextOptions<TeamDbContext> options) : base(options) { }
     
-    public DbSet<Member> Members { get; set; } 
+    public DbSet<User> Members { get; set; } 
     public DbSet<Project> Projects { get; set; }
     
     public DbSet<Team> Teams {  get; set; }
     
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
-        modelBuilder.ApplyConfiguration(new MemberEntityTypeConfiguration());
+        modelBuilder.ApplyConfiguration(new UserEntityTypeConfiguration());
         modelBuilder.ApplyConfiguration(new ProjectEntityTypeConfiguration());
         modelBuilder.ApplyConfiguration(new TeamEntityTypeConfiguration());
         modelBuilder.ApplyConfiguration(new TeamMemberEntityTypeConfiguration());
